fix result of multiplication cast to wider type
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1859595 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwpf/sprm/TableSprmCompressor.java
@@ -81,7 +81,9 @@ public final class TableSprmCompressor
     {
       int itcMac = newTAP.getItcMac();
       byte[] buf = IOUtils.safelyAllocate(
-              1 + (LittleEndian.SHORT_SIZE*(itcMac + 1)) + (TableCellDescriptor.SIZE*itcMac),
+              1
+                + (LittleEndian.SHORT_SIZE*((long)itcMac + 1))
+                + (TableCellDescriptor.SIZE*(long)itcMac),
               MAX_RECORD_LENGTH);
       buf[0] = (byte)itcMac;
